add CardPeek project (card data and info testing)
Cardpeek is a Linux/Windows/Mac OS X tool to read the contents of ISO7816 smart cards. It features a GTK GUI to represent card data in a tree view, and is extendable with a scripting language (LUA).

The goal of this project is to allow smart card owners to be better informed about what type of personal information is stored in these devices.